update mpi_simple test
